# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Система Управления Банковскими Картами - API
  version: 1.0.0
  description: API для создания, блокировки карт, переводов денежных средств и аутентификации.

security:
  - BearerAuth: []

paths:
  /api/v1/auth/login:
    post:
      summary: Аутентификация пользователя и получение JWT-токена
      operationId: login
      tags:
        - Auth Controller
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDto'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Неверный логин или пароль

  /api/v1/cards:
    post:
      summary: Создание новой карты (Только ADMIN)
      operationId: createCard
      tags:
        - Card Controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateDto'
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '400':
          description: Неверные входные данные
        '403':
          description: Доступ запрещен

    get:
      summary: Просмотр списка карт с фильтрацией и пагинацией (USER видит свои, ADMIN — все)
      operationId: getAllCards
      tags:
        - Card Controller
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: cardholder
          in: query
          required: false
          description: Фильтрация по имени владельца (для поиска)
          schema:
            type: string
      responses:
        '200':
          description: Успешное получение списка карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponseDto'
        '403':
          description: Доступ запрещен

  /api/v1/cards/{id}:
    delete:
      summary: Удаление карты (Только ADMIN)
      operationId: deleteCard
      tags:
        - Card Controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта успешно удалена
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена

  /api/v1/cards/{id}/status:
    patch:
      summary: Изменение статуса карты (USER блокирует свою, ADMIN управляет любым статусом)
      operationId: updateCardStatus
      tags:
        - Card Controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [ACTIVE, BLOCKED, EXPIRED]
      responses:
        '200':
          description: Статус карты успешно обновлен
        '400':
          description: Некорректный статус
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена

  /api/v1/cards/transfer:
    post:
      summary: Перевод между своими картами (Только USER)
      operationId: transferBetweenOwnCards
      tags:
        - Card Controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequestDto'
      responses:
        '200':
          description: Перевод успешно выполнен
        '400':
          description: Ошибка валидации или недостаточно средств
        '403':
          description: Доступ запрещен

  /api/v1/admin/users:
    get:
      summary: Получение списка всех пользователей (Только ADMIN)
      operationId: getAllUsers
      tags:
        - Admin User Controller
      security:
        - BearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponseDto'
        '403':
          description: Доступ запрещен

  /api/v1/admin/users/{id}/block:
    patch:
      summary: Блокировка/разблокировка пользователя (Только ADMIN)
      operationId: toggleUserBlock
      tags:
        - Admin User Controller
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Статус пользователя успешно изменен
        '403':
          description: Доступ запрещен
        '404':
          description: Пользователь не найден


components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthRequestDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "admin"

    AuthResponseDto:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."

    CardCreateDto:
      type: object
      required:
        - cardNumber
        - ownerId
        - expiryDate
      properties:
        cardNumber:
          type: string
          pattern: '^\d{16}$'
          example: "1234567812345678"
        ownerId:
          type: integer
          format: int64
          example: 1
        expiryDate:
          type: string
          format: date
          example: "2029-12-31"

    CardResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        cardNumber:
          type: string
          example: "**** **** **** 5678"
        ownerUsername:
          type: string
          example: "ivan_ivanov"
        expiryDate:
          type: string
          format: date
          example: "2029-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          example: "ACTIVE"
        balance:
          type: number
          format: bigdecimal
          example: 1500.50

    CardPageResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponseDto'
        totalElements:
          type: integer
          format: int64
          example: 25
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0

    TransferRequestDto:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
          example: 1
        toCardId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: bigdecimal
          minimum: 0.01
          example: 500.00

    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: "ivan_ivanov"
        role:
          type: string
          example: "ROLE_USER"

    UserPageResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseDto'
        totalElements:
          type: integer
          format: int64
          example: 5
        totalPages:
          type: integer
          example: 1